Adaptive layout for right detail cell
