+++ Stepic/Sources/Frameworks/SettingsTable/Views/Cell/RightDetailCells/Switch/SettingsRightDetailSwitchCellView.swift
