+++ Stepic/Sources/Frameworks/SettingsTable/Views/Cell/RightDetailCells/Switch/SettingsRightDetailSwitchTableViewCell.swift
